requirements.txt
django
boto3
zappa
djangorestframework
django_extensions
django_storages
